transcripts done

+++ DNA_dist_utils.py
@@ -186,7 +186,7 @@ def epoch(model, rank, criterion,
 
 
         with autocast(dtype=torch.bfloat16):
-            encoded_sequence = torch.tensor(tokenizer.encode(data[0]).ids, dtype=torch.long).to(rank).bfloat16() #[:seq_len]
+            encoded_sequence = torch.tensor(tokenizer.encode(data[0]).ids, dtype=torch.long).to(rank) #[:seq_len]
             encoded_sequence = encoded_sequence.unsqueeze(0)
             output = model.forward(encoded_sequence)
             logits, embeddings = output["logits"], output["embeddings"]
